Added pagination to EconItems#GetSchemaItemsForTF2Async() (#105)
* Updated Next property to be of type uint?

If this is too breaking a change, a consideration for `uint` with a default value of 0 would work, but 0 is still a valid value for "start" on the request.

* Updated IEconItems/EconItems#GetSchemaItemsForTF2Async.

added a `uint? start = null` optional property.

* Modified unit tests to take the new Next/start combination.

This verifies that both a) the Next property is valid on each request, and b) that passing `start` along works as intended. This replaces the original test, but could be put alongside it instead.